refactor : ROutineLogSeeder로 분리
